==> globalprep/prs_land-based_nutrient/v2022/STEP1_crop/step4_fubc_merge.Rmd <==
---
title: "OHI 2022: merge FUBC rates with nutrient application dfs"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 1
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '../../../workflow/templates/ohi_hdr.html'
pdf_document:
  toc: true
editor_options: 
  chunk_output_type: console
---

# Summary
Merge the FAOSTAT data with the Bouwman and FUBC datasets

# Preamble
```{r setup, include = FALSE}
# getting packages we want
library(here)
library(janitor)
library(vroom)
library(tidyverse)

source(here('workflow/R/common.R'))

crop_nutrient <- file.path(here(), "globalprep/prs_land-based_nutrient/v2022/STEP1_crop")

region_data()

food_rgns <- read_csv(here("globalprep/spatial/v2021/food_rgns.csv"), col_types = "cdc")

nutrient_d2022_anx <- file.path("/home/shares/ohi/git-annex/globalprep/_raw_data/FAOSTAT/crop_nutrient/d2022")
``` 

# Import

```{r}
fubc_2002 <- vroom(here(crop_nutrient, "int/fubc_2002_gf.csv"))
fubc_2017 <- vroom(here(crop_nutrient, "int/fubc_2017_gf.csv"), 
                   col_select = - fubc_2017_crop)
```

# Checks

```{r}
46 == length(unique(fubc_2002$split_rest_names)) + length(unique(fubc_2017$split_rest_names)) #46
```

# Merge fubc 2002 w/ fubc 2017

```{r}
fubc <- 
  bind_rows(fubc_2002,
            fubc_2017) %>% 
  rename(kg_per_ha = rate) %>% 
  mutate(t_per_ha = kg_per_ha / 1000) %>% # unit conversion
  dplyr::select(- kg_per_ha) %>%
  filter(!is.na(nutrient))

# make sure this is TRUE
244 * 3 * 45 == 
  length(unique(fubc$iso3c)) * # 244 countries
  length(unique(fubc$nutrient)) * # 3 nutrients
  length(unique(fubc$split_rest_names)) # 45 crops
  
244 * 3 * 45 == nrow(fubc)
```

-------------------------------------------------------------------------------

# Add FAO national fertilizer allocation data

```{r}
fao_application <- vroom(here(crop_nutrient, "int/fao_nutrient_application_gf.csv"))

nrow(fao_application) == 244 * 3 * 16
nutrient_df <- left_join(fubc, fao_application)

## should have somewhere around
244*45*16*3  # rows
```

-------------------------------------------------------------------------------

# Add HarvArea column

```{r}
harvested_area_tbl <- 
  vroom(here(crop_nutrient, "data/harvested_area_tbl.csv"),
        col_select = c(iso3c, split_rest_names, harvested_area, prod_system)) %>% 
  filter(prod_system %in% c("H", "I"))

244 * 45 * 2 == nrow(harvested_area_tbl)

nutrient_df <- left_join(nutrient_df, harvested_area_tbl) # good 
```

-------------------------------------------------------------------------------

# Analysis
& total weight is spam_intensity_wt x total_ha x fert_kg_per_ha
&& prop_fert is: total weight divided by the sum of the weights across all crops within a country
```{r}
nutrient_df_full <- 
  nutrient_df %>% 
  mutate(weighting = case_when(prod_system == "I" ~ 2,
                               prod_system == "H" ~ 1,
                               prod_system == "L" ~ 0,
                               prod_system == "S" ~ 0)) %>%
  mutate(total_nutrient = t_per_ha * harvested_area * weighting) %>% 
  
  # Calculate proportion of each type of fertilizer applied to each crop in each country in each year
  
  group_by(iso3c, year, nutrient) %>% 
  mutate(prop_nutrient = total_nutrient / sum(total_nutrient, na.rm = TRUE)) %>% 
  ungroup() %>%
  mutate(prop_nutrient = if_else(is.finite(prop_nutrient) == FALSE,
                                 0,
                                 prop_nutrient)) %>% 
  
  # Allocate FAO national fertilizer application based on the previously calculated proportion
  
  mutate(allocated_nutrient = country_application * prop_nutrient)  %>% 
  mutate(allocated_nutrient = if_else(reported_by_fao == "No",
                                      total_nutrient,
                                      allocated_nutrient)) %>% 
  
  # Relocate
  relocate(iso3c, year, split_rest_names, nutrient, prod_system,
           weighting, t_per_ha, harvested_area, 
           total_nutrient, prop_nutrient,
           country_application, reported_by_fao, allocated_nutrient)

```

# Add gapfilled grassland nutrient to the table
```{r}
fao_nutrient_grassland <- vroom(here(crop_nutrient, "int/fao_nutrient_grassland_1.csv")) %>% 
  filter(reported_by_fao == "Yes") %>% 
  dplyr::select(-reported_by_fao)
  
fubc_grass <- vroom(here(crop_nutrient, "int/fubc_2017_grassland.csv"))

fao_nutrient_grassland_2 <- 
  nutrient_df_full %>% 
  filter(reported_by_fao == "No") %>% 
  left_join(., fubc_grass) %>%
  dplyr::select(iso3c, year, nutrient, allocated_nutrient, grassland) %>% 
  mutate(allocated_grassland_nutrient = allocated_nutrient * grassland) %>% 
  group_by(iso3c, year, nutrient) %>% 
  summarise(country_application = sum(allocated_grassland_nutrient, na.rm = TRUE)) %>% 
  ungroup() %>% 
  bind_rows(fao_nutrient_grassland) 
  
write_csv(fao_nutrient_grassland_2, here(crop_nutrient, "int/fubc_2017_grassland_2.csv"))
```

# Data check

```{r}

nutrient_df_full %>%
  group_by(nutrient, iso3c, year, country_application, reported_by_fao) %>%
  summarise(total = sum(prop_nutrient)) %>%
  view()

nrow(nutrient_df_full) == 244 * 2 * 3 * 45*16

sum(nutrient_df_full$total_nutrient)
sum(unique(nutrient_df_full$country_application), na.rm = TRUE)
sum(nutrient_df_full$allocated_nutrient, na.rm = TRUE)

nutrient_df_full %>% 
  group_by(reported_by_fao) %>% 
  summarise(total_nutrient = sum(total_nutrient, na.rm = TRUE),
            harvested_area = sum(harvested_area, na.rm = TRUE)) %>% 
  ungroup() %>% 
  mutate(rate = total_nutrient / harvested_area)
  

nutrient_df_full %>% 
  filter(iso3c == "ALA") %>% 
  group_by(nutrient) %>% 
  summarise(sum(allocated_nutrient)) %>% 
  arrange(nutrient
  )


tbl1 <- 
  nutrient_df_full %>% 
  group_by(nutrient, year, iso3c) %>% 
  summarise(allocated = sum(allocated_nutrient)) %>% 
  ungroup() %>% 
  arrange(nutrient)

nutrient_df_full %>% 
  group_by(nutrient, year, iso3c) %>% 
  dplyr::select(nutrient, year, country_application) %>% 
  unique() %>% 
  arrange(nutrient) %>%
  left_join(tbl1) %>% 
  mutate(diff = country_application / allocated) %>% 
  view()

write_csv(nutrient_df_full, here(crop_nutrient, "int/nutrient_df_full.csv"))
```

# Trim
```{r}
nutrient_df <- 
  nutrient_df_full %>% 
  dplyr::select(iso3c, year, nutrient, prod_system, split_rest_names, allocated_nutrient)
```

# Save
```{r}
write_csv(nutrient_df, here(crop_nutrient, "int/nutrient_df.csv"))
```
