==> globalprep/prs_land-based_nutrient/v2022/STEP1_crop/step5_crop_nutrient_mapping.Rmd <==
---
title: "OHI 2022: crop_nutrient_mapping"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 1
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '../../../workflow/templates/ohi_hdr.html'
pdf_document:
  toc: true
editor_options: 
  chunk_output_type: console
---

# Summary
Map yearly nutrient leaching from crops using static crop farming location rasters taken from the Global Food Project (Halpern et al. 2021 in review): https://github.com/OHI-Science/food_systems/

# Preamble
```{r setup, include = FALSE}
# getting packages we want
library(here);library(janitor);library("future");library(furrr);library(vroom);library(raster); library(tidyverse); library(terra)

# Raster templates

# Import food_rgns xy df
food_rgns_xy <-
  read.csv(file.path("/home/shares/ohi/git-annex/globalprep/spatial/v2021/food_rgns_xy.csv")) %>%
  dplyr::select(x, y, iso3c)

food_crs <- "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"


source(here('workflow/R/common.R'))

crop_nutrient <- file.path(here(), "globalprep/prs_land-based_nutrient/v2022/STEP1_crop")

region_data()

food_rgns <- read_csv(here("globalprep/spatial/v2021/food_rgns.csv"), col_types = "cdc")

nutrient_d2022_anx <- file.path("/home/shares/ohi/git-annex/globalprep/_raw_data/FAOSTAT/crop_nutrient/d2022")

# Import main dataframe
nutrient_df <- vroom::vroom(here(crop_nutrient, "int/nutrient_df.csv"))
distinct_vars <- expand.grid(unique(nutrient_df$split_rest_names),
                             unique(nutrient_df$nutrient), 
                             2005:2020)

prep <- file.path("/home/shares/ohi/git-annex/globalprep/prs_land-based_nutrient/v2022/")

food_systems_crop_areas <- file.path("/home/shares/ohi/git-annex/globalprep/_raw_data/halpern_food_systems/d2021/crop/cell_allocation/")
``` 

# Withdrawal of N

Make a raster that describes proportion withdrawal by plants (Table V, NUE values, Bouwman et al. 2005). NOTE: When NUE > 1, use proportion 1. Multiply this raster by the tonnes N applied. Save this raster in dataprep. Check: The sum of the entire raster should be somewhere around 50% of total global fertilizer application.

```{r}
N_withdrawal_raster <- 
  read.csv(here(crop_nutrient, "raw/bouwman_2005_tbl5.csv"))[,c(1,4)] %>% 
  mutate(NUE = if_else(NUE >= 100, 1, NUE / 100)) %>% 
  left_join(food_rgns_xy, ., by = "iso3c") %>% 
  dplyr::select(x, y, NUE) %>%
  #as.matrix() %>%
  rasterFromXYZ(crs = food_crs)

plot(N_withdrawal_raster)
```

# Volatilization of N as NH3
Make a raster that describes proportion of N lost to volatilization as NH3. This is based on Table 4 of Bouwman et al. 2002. Ultimately, I decided the best bet is to add the grassland, flooded rice, and crops to get the average values based on how the fertilizer is regionally applied (so this will be slightly different from what you created already). I have done the calculations in the attached spreadsheet (this column will create the raster: prop_NH3-N total). Multiply this raster by the tonnes N applied. Save this raster in the predata folder (people will want this one). Check: The sum of the entire raster should be somewhere around 14% of total global fertilizer application.

```{r}
N_volatilized_raster <- 
  read.csv(here(crop_nutrient, "raw/bouwman_2002_tbl4.csv")) %>%
  dplyr::select(prop_nh3_n_total, iso3c) %>% 
  left_join(food_rgns_xy, ., by = "iso3c") %>% 
  dplyr::select(x, y, prop_nh3_n_total) %>% 
  # as.matrix() %>%
  rasterFromXYZ(crs = food_crs)
plot(N_volatilized_raster)
```

# Map excess fertilizer applications  (similar to FAO estimates for manure)

```{r}
nutrient_application <- function(this_crop, this_nutrient, this_year) {

  # this_crop = "acof"
  # this_nutrient = "N"
  # this_year = 2010

  # Filter by crop and nutrient for each production system
  raster_H <- 
    nutrient_df %>% 
    filter(split_rest_names %in% this_crop &
           nutrient %in% this_nutrient &
           prod_system %in% "H",
           year == this_year) %>%
    dplyr::select(iso3c, year, allocated_nutrient) %>% 
    left_join(food_rgns_xy, ., by = "iso3c") %>% # Add to XY dataframe
    dplyr::select(-iso3c) %>% 
    raster::rasterFromXYZ(crs = food_crs)
  
  raster_H <- raster_H[[2]] # filter for values 

  raster_I <- 
    nutrient_df %>% 
    filter(split_rest_names %in% this_crop &
           nutrient %in% this_nutrient &
           prod_system %in% "I", 
           year == this_year) %>%
    dplyr::select(iso3c, year, allocated_nutrient) %>% 
    left_join(food_rgns_xy, ., by = "iso3c") %>% 
    dplyr::select(-iso3c) %>% 
    raster::rasterFromXYZ(crs = food_crs)
  
    raster_I <- raster_I[[2]] # filter for values 


  # Overlay to calculate total nutrient application at the cell-level
    
  nutrient_application <- raster::overlay(
    
   raster_H, # High input layer
    raster::raster(file.path(food_systems_crop_areas,
                     "crop_",
                     this_crop,
                     "_H_harvested_area_allocation.tif", 
                     fsep = "")),
    
    raster_I, # Irrigation layer
    raster::raster(file.path(food_systems_crop_areas,
                     "crop_",
                     this_crop,
                     "_I_harvested_area_allocation.tif", 
                     fsep = "")),
    
    fun = function(r1, r2, r3, r4) {
      
      r1[is.na(r1)] <- 0
      r2[is.na(r2)] <- 0
      r3[is.na(r3)] <- 0
      r4[is.na(r4)] <- 0
      
      r5 <- r1 * r2
      r6 <- r3 * r4

      return(r5 + r6)
      
      })
  
  # Save nutrient application
  raster::writeRaster(nutrient_application, 
              file.path(prep,
                        "crop_int/crop_",
                        this_crop,
                        "_application_",
                        this_nutrient,
                        "_",
                        this_year,
                        ".tif", 
                        fsep = ""), 
              format = "GTiff", 
              overwrite = TRUE)
  
  if (this_nutrient == "N") {
  
  # Withdrawal of N
  
  N_withdrawal_raster_outcome <- 
    raster::overlay(N_withdrawal_raster,
            nutrient_application,
            fun = function(r1, r2){
              
              r1[is.na(r1)] <- 0
              r2[is.na(r2)] <- 0
              
              r1 * r2
              
            })
    
    raster::writeRaster(N_withdrawal_raster_outcome,
                file.path(prep,
                          "crop_int/crop_",
                          this_crop,
                          "_withdrawal_N_",
                          this_year, 
                          ".tif", 
                          fsep = ""), 
                format = "GTiff", 
                overwrite = TRUE)
  
  # Volatilization of N as NH3
  
    N_volatilized_raster_outcome <- 
      raster::overlay(N_volatilized_raster,
              nutrient_application,
              fun = function(r1, r2){
                
                r1[is.na(r1)] <- 0
                r2[is.na(r2)] <- 0
                
                r1 * r2
              })
      
      raster::writeRaster(N_volatilized_raster_outcome,
                  file.path(prep,
                          "crop_predatalayers/crop_",
                          this_crop,
                          "_volatilized_N_",
                          this_year,
                          ".tif", 
                          fsep = ""), 
                  format = "GTiff", 
                  overwrite = TRUE) 

    
    # Denitrification, N2 and N20 emissions
    ## We are ignoring NO emissions, which are about 0.7% of N application.
    ## Multiply the tonnes applied raster by 0.01 and then by 10.2 to get the 
    ## amount of N lost to denitrification. 
    
    N_denitrification_raster <- 
      nutrient_application *
      0.01 *
      10.2
        
    raster::writeRaster(N_denitrification_raster,
                file.path(prep,
                          "crop_int/crop_",
                          this_crop,
                          "_denitrification_N_",
                          this_year,
                          ".tif", 
                          fsep = ""), 
                format = "GTiff", 
                overwrite = TRUE)
      
   # Leaching/runoff
   ## Sum the tonnes N rasters for: withdrawal, volatilization, and denitrification.
   ## Leaching/runoff: Tonnes applied minus N - (withdrawal + volatilization + denitrification)
   ## NOTE: when leaching/runoff < 0, make 0
   ## Save this raster in predata folder. Report on global average of leaching/runoff
      
   N_leaching <- 
      raster::overlay(nutrient_application,
              N_withdrawal_raster_outcome,
              N_volatilized_raster_outcome,
              N_denitrification_raster,
              fun = function(applied,
                             withdrawn,
                             volatilized,
                             denitrified) { 
                
                leached <- applied - (withdrawn + volatilized + denitrified)
                leached[leached < 0] <- 0
                return(leached)
                
                })
   
    names(N_leaching) <- paste("land_", this_crop, "_nutrient_tonnes_N", sep = "")
     
    raster::writeRaster(N_leaching,
                file.path(prep,
                          "crop_predatalayers/",
                          this_crop,
                          "_leached_N_",
                          this_year,
                          fsep = ""), 
                format = "GTiff", 
                overwrite = TRUE)
    
  } else if (this_nutrient == "P2O5") {
    
    leaching_raster_P <-
      nutrient_application /
      2.29 * # Convert to P
      0.065 # Leaching proportion
      
  leaching_raster_P[is.na(leaching_raster_P)] <- 0
    
  names(leaching_raster_P) <- 
    paste("land_", 
          this_crop, 
          "_nutrient_tonnes_P", 
          sep = "")

  raster::writeRaster(leaching_raster_P,
              file.path(prep,
                        "crop_predatalayers/",
                          this_crop,
                          "_leached_P_",
                        this_year,
                          fsep = ""),
              format = "GTiff",
              overwrite = TRUE)
  }
 }
plan(multisession, workers = 12)

options <- furrr_options(seed = 54232)

future_pmap(list(distinct_vars$Var1,
            distinct_vars$Var2,
            distinct_vars$Var3),
            .f = nutrient_application,
            .options = options)

```

# Calculate nutrient leaching

```{r}
## will need to add year argument here; stack and sum by year all volatilized and leached; per nutrient? 

total_leaching <- function(crop, this_year) {
  #  crop = "acof"
  # this_year = 2010


  leaching_raster <-
    list.files(file.path(prep, "crop_predatalayers"),
               pattern = paste(crop, "_(leached|volatilized)",  "_(N|P)_", this_year, sep = ""),
               full.names = TRUE) %>%
    terra::rast() %>%
    terra::tapp(., 1, fun = "sum")

  names(leaching_raster) <- paste("land_", crop, "_nutrient_tonnes",   sep = "")

  # Save as Raster in /dataprep/ folder
  writeRaster(leaching_raster,
              file.path(prep,
                        "crop_stressors/land_",
                        crop,
                        "_crop_produce_nutrient_",
                        this_year,
                        ".tif",
                        fsep = ""),
              overwrite = TRUE)

}

plan(multisession, workers = 12)

options <- furrr_options(seed = 54232)

future_pmap(list(distinct_vars$Var1,
            distinct_vars$Var3),
            .f = total_leaching,
            .options = options)

## check how many files
n_years = 16 
n_crops = 45

n_years*n_crops == length(list.files("/home/shares/ohi/git-annex/globalprep/prs_land-based_nutrient/v2022/crop_stressors")) # TRUE
```



# Test ouputs
```{r}
nutrient_test <- rast(file.path(prep, "crop_int/crop_tnut_application_K2O_2018.tif"))
plot(log(nutrient_test+1))
global(nutrient_test, "sum") # 626168.2
nutrient_df %>% 
  filter(split_rest_names %in% "tnut",
         nutrient == "K2O") %>% 
  summarise(sum(allocated_nutrient, na.rm = TRUE)) # 8215343

```

# Check that withdrawal is 50% of total
```{r}
total_N <- 
  list.files(file.path(prep,
                     "crop_int",
                     fsep = ""),
             pattern = "_application_N_2017",
             full.names = TRUE) %>% 
  .[!grepl(c("toba|teas|cof|othr"), .)] %>% 
  rast() %>% 
  tapp(., 1, fun = "sum") %>% 
  global(., "sum") # 103091247

withdrawn_N <- 
  list.files(file.path(prep,
                     "crop_int",
                     fsep = ""),
             pattern = "withdrawal_N_2019",
             full.names = TRUE) %>% 
  .[!grepl(c("toba|teas|cof|othr"), .)] %>% 
  rast() %>% 
  tapp(., 1, fun = "sum") %>% 
  global(., "sum")

withdrawn_N / total_N # should be approx 50% of total N ; 0.5656585 (2017)

# NH3 volatilization
volatilized_N <- 
  list.files(file.path(prep,
                     "crop_predatalayers",
                     fsep = ""),
             pattern = "volatilized_N_2017",
             full.names = TRUE) %>% 
  .[!grepl(c("toba|teas|cof|othr"), .)] %>% 
  rast() %>% 
  tapp(., 1, fun = "sum") %>% 
  global(., "sum")

volatilized_N / total_N # should be approx 14% of total N; 0.1553757 (2017)

# Denitrification, N2 and N20 emissions
denitrification_N <- 
  list.files(file.path(prep,
                     "crop_int/",
                     fsep = ""),
             pattern = "denitrification_N_2017",
             full.names = TRUE) %>% 
  .[!grepl(c("toba|teas|cof|othr"), .)] %>% 
  rast() %>% 
  tapp(., 1, fun = "sum") %>% 
  global(., "sum")

denitrification_N / total_N # should be approx 10% of total N; 0.102 (2017)

leaching_N <- 
  list.files(file.path(prep,
                       "crop_predatalayers/",
                       fsep = ""),
             pattern = "leached_N_2017",
             full.names = TRUE) %>% 
  .[!grepl(c("toba|teas|cof|othr"), .)] %>% 
  rast() %>% 
  tapp(., 1, fun = "sum") %>% 
  global(., "sum")
  
leaching_N / total_N # 0.1872403 (2017)

# Check P
P_leached <- 
  list.files(file.path(prep, "crop_predatalayers/"),
             pattern = "_leached_P_2017",
             full.names = TRUE) %>% 
  .[!grepl(c("toba|teas|cof|othr"), .)] %>% 
  rast()  %>% 
  tapp(., 1, fun = "sum") %>% 
  global(., fun = "sum")
           
total_P <- 
  list.files(file.path(prep,
                     "crop_int/",
                     fsep = ""),
             pattern = "_application_P2O5_2017",
             full.names = TRUE) %>% 
  .[!grepl(c("toba|teas|cof|othr"), .)] %>% 
  rast() %>% 
  tapp(., 1, fun = "sum") %>% 
  global(., "sum")

(total_P / 2.29) * 0.065  # 1198960
P_leached # 1198960 ; perfect

# Check final values
total_leached <- 
  list.files(file.path(prep, "crop_stressors/"),
             pattern = "land_.*_crop_produce_.*nutrient_2017",
             full.names = TRUE) %>% 
  .[!grepl(c("toba|teas|cof|othr"), .)] %>% 
  rast() %>% 
  tapp(., 1, fun = "sum") %>% 
  global(., "sum")

leaching_N + P_leached + volatilized_N # 36519680
total_leached # 36519680 ; perfect

fodd_applications <- 
  list.files(file.path(prep,
                     "crop_int/",
                     fsep = ""),
             pattern = "crop_fodd_application_N_2017",
             full.names = TRUE) %>% 
  rast() %>% 
  tapp(., 1, fun = "sum") %>% 
  global(., "sum")
fodd_applications / total_N # 0.05171939
```
